requirements.txt
asgiref==3.8.1
attrs==24.3.0
boto3==1.35.97
botocore==1.35.97
certifi==2024.12.14
charset-normalizer==3.4.1
decorator==5.1.1
Django==5.1.4
django-cleanup==9.0.0
django-common-helpers==0.9.2
django-cors-headers==4.6.0
django-environ==0.11.2
django-model-utils==5.0.0
django-mysql==4.15.0
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
fonttools==4.55.3
freezegun==1.5.1
future==1.0.0
html5lib==1.1
idna==3.10
jmespath==1.0.1
joblib==1.4.2
mysqlclient==2.2.7
pillow==11.1.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
s3transfer==0.10.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
validators==0.34.0
webencodings==0.5.1
